Add controller test helpers to mock Auth0
I don't want to pass around actual JWT tokens for most of these
controller tests, so these helpers will inject a mock authenticator that
accepts any token and returns a test-specified auth0_id.

The auth0_headers helper is mostly for convenience. We could pass any
token, but we have to pass *some* token, so it fixes a fake value and
builds the Authorization header.