Upgrade to latest JSONAPI:Resources and branch of JSONAPI:Authorization
As well as the latest patch of Rails. There's a bug in the way that
JSONAPI:Resources handles JSONAPI:Authorization's Pundit scopes with
associations -- it builds incorrect SQL and looks associated resources up by the
wrong ids.

Unfortunately, upgrading doesn't actually fix the issue, but it doesn't hurt and
it's nice to be on the latest. This does require using a branch version of
JSONAPI:Authorization. We might want to revisit that before doing anything real.