Rewrite Aspect.{character,world} logic using self-reference
The logic I had wasn't working -- or at least wasn't working in Rails 7, but I
suspect it never really worked.

The new logic uses a self-reference to link to characters and worlds. As
explained in the comment, we can't assume aspects is in the FROM list of the
query, but we need it to get aspected_type.