Create a concern for including attachments in resources
This concern will create all the necessary attributes and methods to make
attachments and variants appear as links on a resource, and to allow attaching
via a signed id from direct upload.

Note: This only supports a single attachment. Handling multiple attachments will
be a future problem if and when we need it.